split point, line, and polygon classes
