+++ src/libs2-s2geography.cpp
@@ -12,6 +12,9 @@
 #include "wk/geometry-handler.h"
 
 #include "libs2-geography.h"
+#include "libs2-point-geography.h"
+#include "libs2-polyline-geography.h"
+#include "libs2-polygon-geography.h"
 #include <Rcpp.h>
 using namespace Rcpp;
